refactor: extract validation functions to separate module
- Created new validator.js module to contain all form validation logic
- Updated imports in formHandler.js to use the extracted validation functions
- Updated test file to import validation functions from the new module
- All tests pass with the new modular structure
